requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-openai==0.3.32
langchain-chroma==0.2.5
chromadb==1.0.20
python-dotenv==1.1.1
tqdm==4.67.1
rank-bm25==0.2.2
